refactor(loop): reorganize package into multiple files
Extract git operations to git.go and prompt building to prompt.go.
Move Mode and Config types to types.go. Rename provider.go to providers.go.
