+++ internal/loop/git.go
@@ -1,0 +1,37 @@
+package loop
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+	"strings"
+)
+
+// getCurrentBranch returns the current git branch name
+func getCurrentBranch() (string, error) {
+	cmd := exec.Command("git", "branch", "--show-current")
+	output, err := cmd.Output()
+	if err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(string(output)), nil
+}
+
+// pushChanges pushes commits to the remote branch
+func pushChanges(branch string) error {
+	// Try to push
+	cmd := exec.Command("git", "push", "origin", branch)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Run(); err != nil {
+		// If push failed, try to create remote branch
+		fmt.Println("Failed to push. Creating remote branch...")
+		cmd = exec.Command("git", "push", "-u", "origin", branch)
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		return cmd.Run()
+	}
+
+	return nil
+}
